Move out AssignmentProcessor and fix TheoryOfHeapFunApp
